Make the malicious operator and test domain node starting process same as production
